implement the WalkingModule application
