+++ cmake/WalkingControllersFindDependencies.cmake
@@ -159,3 +159,6 @@ walking_controllers_dependent_option(WALKING_CONTROLLERS_COMPILE_libKinDyn_wrapp
 walking_controllers_dependent_option(WALKING_CONTROLLERS_COMPILE_libRetargeting_helper "Compile Retargeting_helper?" ON
                                     "WALKING_CONTROLLERS_HAS_iDynTree;WALKING_CONTROLLERS_COMPILE_libYARP_helper;WALKING_CONTROLLERS_HAS_ICUB" OFF)
 walking_controllers_dependent_option(WALKING_CONTROLLERS_COMPILE_libLogger_client "Compile Logger_client?" ON WALKING_CONTROLLERS_COMPILE_libYARP_helper OFF)
+
+walking_controllers_dependent_option(WALKING_CONTROLLERS_COMPILE_WalkingModule "Compile WalkingModule app?" ON
+  "WALKING_CONTROLLERS_COMPILE_libYARP_helper;WALKING_CONTROLLERS_COMPILE_libiDynTree_helper;WALKING_CONTROLLERS_COMPILE_libRobot_helper;WALKING_CONTROLLERS_COMPILE_libKinDyn_wrapper;WALKING_CONTROLLERS_COMPILE_libTrajectory_planner;WALKING_CONTROLLERS_COMPILE_libSimplified_model_controllers;WALKING_CONTROLLERS_COMPILE_libWhole_body_controllers;WALKING_CONTROLLERS_COMPILE_libRetargeting_helper;WALKING_CONTROLLERS_COMPILE_libLogger_client" OFF)
