+++ src/CMakeLists.txt
@@ -13,3 +13,4 @@ add_subdirectory(libTrajectory_planner)
 add_subdirectory(libKinDyn_wrapper)
 add_subdirectory(libRetargeting_helper)
 add_subdirectory(libLogger_client)
+add_subdirectory(WalkingModule)
