+++ src/WalkingModule/thrifts/WalkingCommands.thrift
